Renamed all the folders to be the correct spelling of 'Mathematic' instead of whatever stupid spelling I did
