+++ app/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="swen_anigans.mathmaticfanatic">
+    package="swen_anigans.mathematicfanatic">
 
     <application
         android:allowBackup="true"
