+++ app/src/main/res/layout/activity_quiz_start.xml
@@ -8,7 +8,7 @@
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
     android:paddingBottom="@dimen/activity_vertical_margin"
-    tools:context="swen_anigans.mathmaticfanatic.QuizStartActivity">
+    tools:context=".QuizStartActivity">
 
     <TextView
         android:layout_width="wrap_content"
